requirements.txt
-i https://pypi.org/simple
asgiref==3.7.2 ; python_version >= '3.7'
certifi==2023.5.7 ; python_version >= '3.6'
cffi==1.15.1
charset-normalizer==3.1.0 ; python_full_version >= '3.7.0'
cryptography==41.0.1
defusedxml==0.7.1 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
distlib==0.3.6
dj-database-url==2.0.0
django==4.2.2
django-allauth==0.54.0
django-contact-form==2.0.1
django-cors-headers==4.1.0
django-crispy-forms==2.0
django-environ==0.10.0
django-filter==23.2
djangorestframework==3.14.0
environ==1.0
filelock==3.12.2 ; python_version >= '3.7'
idna==3.4 ; python_version >= '3.5'
markdown==3.4.3
oauthlib==3.2.2 ; python_version >= '3.6'
pillow==9.5.0
pipenv==2023.6.18
platformdirs==3.8.0 ; python_version >= '3.7'
psycopg2-binary==2.9.6
pycparser==2.21
pyjwt[crypto]==2.7.0 ; python_version >= '3.7'
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3
requests==2.31.0 ; python_version >= '3.7'
requests-oauthlib==1.3.1 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
setuptools==68.0.0 ; python_version >= '3.7'
sqlparse==0.4.4 ; python_version >= '3.5'
typing-extensions==4.6.3
urllib3==2.0.3 ; python_version >= '3.7'
virtualenv==20.23.1 ; python_version >= '3.7'
virtualenv-clone==0.5.7 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
